refactor(markdown-store): remove old store pattern (Observable based)
