+++ src/app/app.module.config.ts
@@ -7,9 +7,6 @@ import { CodeHighlighter } from '@app/shared/services/code-highlighter/code-high
 import { HighlightJs } from '@app/shared/services/code-highlighter/highlight-js';
 import { CustomSanitizer } from '@app/shared/services/custom-sanitizer/custom-sanitizer';
 import { DomPurify } from '@app/shared/services/custom-sanitizer/dom-purify';
-import { MarkdownStore } from './shared/services/store/markdown.store';
-import { MarkdownState } from './shared/services/store/markdown-state.model';
-import { Toolbar } from './shared/enums/toolbars.enum';
 import { NotesService } from './pages/notes/notes.service';
 import { StorageStrategy } from './shared/services/storage-strategy/storage-strategy';
 import { LocalStorageStrategy } from './shared/services/storage-strategy/local-storage-strategy';
@@ -21,19 +18,6 @@ export const markedFactory = (
   customSaniziter: CustomSanitizer,
 ): Marked => {
   return new Marked(sanitizer, codeHighlighter, customSaniziter);
-};
-
-export const markdownStoreFactory = (): MarkdownStore => {
-  const initialState: MarkdownState = {
-    markdownText: '',
-    htmlText: '',
-    checked: {
-      [Toolbar.Preview]: true,
-    },
-    currentActiveNote: null,
-    notes: [],
-  };
-  return new MarkdownStore(initialState);
 };
 
 const notesServiceFactory = (...storageStrategies: StorageStrategy[]) => {
@@ -55,11 +39,6 @@ export const codeHighlighterProvider: Provider = {
 export const customSanitizerProvider: Provider = {
   provide: CustomSanitizer,
   useClass: DomPurify,
-};
-
-export const markdownStoreProvider: Provider = {
-  provide: MarkdownStore,
-  useFactory: markdownStoreFactory,
 };
 
 export const notesServiceProvider: Provider = {
