+++ src/app/pages/notes/note-details/note-details.component.ts
@@ -2,7 +2,7 @@ import { Component, ViewEncapsulation, OnInit } from '@angular/core';
 import { Store, select } from '@ngrx/store';
 import { Observable } from 'rxjs';
 
-import { INote } from '@app/shared/services/store/markdown-state.model';
+import { INote } from '@app/shared/models/markdown-state.model';
 import * as fromRoot from '@app/pages/notes/shared/reducers';
 import { NotesActions } from '../shared/actions';
