+++ src/app/pages/notes/notes.service.ts
@@ -2,8 +2,7 @@ import { Injectable } from '@angular/core';
 import { Observable, of } from 'rxjs';
 import { delay } from 'rxjs/operators';
 
-import { INote } from '@app/shared/services/store/markdown-state.model';
-import { generateRandomId } from '@app/shared/utils/generator.util';
+import { INote } from '@app/shared/models/markdown-state.model';
 import { StorageStrategy } from '@app/shared/services/storage-strategy/storage-strategy';
 import { EStorageStrategy } from '@app/shared/enums/strategy.enum';
