+++ src/app/pages/notes/shared/actions/navigations.actions.ts
@@ -1,7 +1,7 @@
 import { createAction, props } from '@ngrx/store';
 import { Update } from '@ngrx/entity';
 
-import { INote } from '@app/shared/services/store/markdown-state.model';
+import { INote } from '@app/shared/models/markdown-state.model';
 
 export const clickNote = createAction(
   'CLICK_NOTE',
