+++ src/app/shared/constants/note.const.ts
@@ -1,4 +1,4 @@
-import { INote } from '@app/shared/services/store/markdown-state.model';
+import { INote } from '@app/shared/models/markdown-state.model';
 import { generateRandomId } from '@app/shared/utils/generator.util';
 
 export const createDefaultNote = (): INote => ({
