+++ src/app/shared/models/markdown-state.model.ts
