--- src/app/shared/services/store/store.ts
@@ -1,20 +1,0 @@
-import { Observable, BehaviorSubject } from 'rxjs';
-
-export class Store<T> {
-  state$: Observable<T>;
-
-  private dispatcher: BehaviorSubject<T>;
-
-  constructor(initialState: T) {
-    this.dispatcher = new BehaviorSubject(initialState);
-    this.state$ = this.dispatcher.asObservable();
-  }
-
-  get state(): T {
-    return this.dispatcher.getValue();
-  }
-
-  protected setState(nextState: T): void {
-    this.dispatcher.next(nextState);
-  }
-}
