+++ src/app/shared/utils/sort.util.ts
@@ -1,4 +1,4 @@
-import { INote } from '@app/shared/services/store/markdown-state.model';
+import { INote } from '@app/shared/models/markdown-state.model';
 
 export function sortDescendingByDateCreated(a: INote, b: INote): number {
   return b.dateCreated - a.dateCreated;
